Add System probe.

+++ QueryGlass/src/Domain/Entities/CpuDetail.cs
@@ -3,7 +3,7 @@
 public class CpuDetail
 {
     public int CpuCores { get; set; }
-    public int CpuCoreUsage { get; set; }
+    public double CpuCoreUsage { get; set; }
     public int CpuProcessCount { get; set; }
     public int CpuThreadCount { get; set; }
 }
